requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
fastapi==0.135.1
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
h11==0.16.0
idna==3.11
neo4j==6.1.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2025.2
six==1.17.0
starlette==0.52.1
starlette-graphene==0.4.0
starlette-graphene3==0.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
